Fixed openssl signature bug

+++ src/structures/openssl.rs
@@ -7,7 +7,7 @@ pub struct OpenSSLCryptHeader {
 
 /// Parse an OpenSSl crypto header
 pub fn parse_openssl_crypt_header(ssl_data: &[u8]) -> Result<OpenSSLCryptHeader, StructureError> {
-    let ssl_structure = vec![("magic", "u32"), ("salt", "u64")];
+    let ssl_structure = vec![("magic", "u64"), ("salt", "u64")];
 
     if let Ok(ssl_header) = common::parse(ssl_data, &ssl_structure, "big") {
         return Ok(OpenSSLCryptHeader {
